refactor(rpc): extract jsonrpc helpers to submodule (backlog #11, phase 1)
Move jsonrpc.rs → jsonrpc/mod.rs and extract all shared helpers
(to_hex, normalize_rpc_*, parse_hex_u64, resolve_block_tag,
parse_address_filter, parse_topic_filter, log_matches, collect_logs,
load_logs_for_tx, block_gas_used_ratio, TopicFilter) into
jsonrpc/helpers.rs. The dispatcher, handler match arms, and
JSON-RPC envelope types stay in mod.rs unchanged.

No behaviour change. mod.rs shrinks 1302 → 1057 lines; helpers.rs is
233 lines. All tests pass, clippy clean.

Phase 2 (namespace split — eth/net/web3/sentrix each in their own
module with per-namespace dispatch) remains pending as the bigger
mechanical extraction.
